Tests/LibWeb/Text/expected/HTML/set-innerText.txt

Empty string: ""
Simple text: "foo"
Text with newlines: "<br>foo<br>bar<br>baz<br><br>foobar<br>"
